wip: Reorganizing code
