+++ MiAZ/frontend/desktop/widgets/sidebar.py
@@ -58,10 +58,6 @@ class MiAZSidebar(Adw.Bin):
         # Setup system menu
         self._setup_menu_system()
 
-        # Dropdown filters
-        self.toolbar_filters = self._setup_toolbar_filters()
-        self.app.add_widget('workspace-toolbar-filters', self.toolbar_filters)
-
         # Status page
         self.status_page = Adw.StatusPage(
             title=_(""),
@@ -72,6 +68,7 @@ class MiAZSidebar(Adw.Bin):
         )
 
         menubutton_system = self.app.get_widget('headerbar-button-menu-system')
+        toolbar_filters = self.app.get_widget('workspace-toolbar-filters')
 
         self.set_child(
             MiAZToolbarView(
@@ -91,7 +88,7 @@ class MiAZSidebar(Adw.Bin):
                         # ~ Gtk.ScrolledWindow(
                             # ~ propagate_natural_height=True, child=Gtk.Box()
                         # ~ ),
-                        self.toolbar_filters,
+                        toolbar_filters,
                         self.status_page,
                     ],
                 ),
@@ -127,44 +124,6 @@ class MiAZSidebar(Adw.Bin):
         menubutton.set_popover(popover=popover)
         self.app.add_widget('headerbar-button-menu-system', menubutton)
 
-    def _setup_toolbar_filters(self):
-        factory = self.app.get_service('factory')
-        dropdowns = self.app.get_widget('ws-dropdowns')
-        widget = factory.create_box_vertical(spacing=0, margin=0, hexpand=True, vexpand=False)
-        body = factory.create_box_vertical(margin=3, spacing=6, hexpand=True, vexpand=True)
-        body.set_margin_top(margin=6)
-        body.set_margin_start(margin=12)
-        body.set_margin_end(margin=12)
-        widget.append(body)
-        row_up = factory.create_box_vertical(margin=3, spacing=6, hexpand=True, vexpand=True)
-        row_down = factory.create_box_vertical(margin=3, spacing=6, hexpand=True, vexpand=True)
-        body.append(row_up)
-        body.append(row_down)
-        widget.append(Gtk.Separator.new(orientation=Gtk.Orientation.HORIZONTAL))
-
-        dropdowns = {}
-        ### Projects dropdown
-        i_type = Project.__gtype_name__
-        i_title = _(Project.__title__)
-        dd_prj = factory.create_dropdown_generic(item_type=Project)
-        boxDropdown = factory.create_box_filter(i_title, dd_prj)
-        dropdowns[i_type] = dd_prj
-        row_up.append(boxDropdown)
-
-        for item_type in [Country, Group, SentBy, Purpose, SentTo]:
-            i_type = item_type.__gtype_name__
-            i_title = _(item_type.__title__)
-            dropdown = factory.create_dropdown_generic(item_type=item_type)
-            boxDropdown = factory.create_box_filter(i_title, dropdown)
-            row_down.append(boxDropdown)
-            dropdowns[i_type] = dropdown
-
-        self.app.add_widget('ws-dropdowns', dropdowns)
-        # ~ btnClearFilters = factory.create_button(icon_name='io.github.t00m.MiAZ-entry_clear', tooltip='Clear all filters', css_classes=['flat'], callback=self.clear_filters)
-        # ~ boxDropdown = factory.create_box_filter('', btnClearFilters)
-        # ~ row_up.append(boxDropdown)
-
-        return widget
 
     def clear_filters(self, *args):
         # ~ search_entry = self.app.get_widget('searchentry')
